#26 move jwt fetch logic to shared class
