+++ src/test/kotlin/com/example/householdExpenses/integration/ExpenseIntegrationTests.kt
@@ -1,5 +1,6 @@
 package com.example.householdExpenses.integration
 
+import com.example.householdExpenses.IntegrationTestUtils
 import com.example.householdExpenses.core.security.SecurityConfig
 import com.example.householdExpenses.domain.Fixtures
 import org.junit.jupiter.api.Assertions.*
@@ -12,7 +13,6 @@ import org.springframework.http.MediaType
 import org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.get
-import org.springframework.test.web.servlet.post
 
 /**
  * [com.example.householdExpenses.presentation.api.expense.ExpenseController]
@@ -22,22 +22,12 @@ import org.springframework.test.web.servlet.post
 @AutoConfigureMockMvc
 @Import(SecurityConfig::class)
 internal class ExpenseIntegrationTests @Autowired constructor(
-    val mockMvc: MockMvc,
-) {
+    val mockMvc: MockMvc
+){
     val requestPath = "/api/expenses";
 
-    private fun fetchJwt(): String {
-        val tokenRequest = mockMvc.post("/api/token") {
-            with(SecurityMockMvcRequestPostProcessors.httpBasic("user1@example.com", "1qazxsw2"))
-        }
-            .andExpect { status { isOk() } }
-            .andReturn()
-
-        return tokenRequest.response.contentAsString
-    }
-
     @Test
-    internal fun `get expenses with HttpBasicAuth`() {
+    internal fun `get expenses with Basic Authentication`() {
         mockMvc.get(requestPath) {
             with(SecurityMockMvcRequestPostProcessors.httpBasic("user1@example.com", "1qazxsw2"))
         }
@@ -61,8 +51,8 @@ internal class ExpenseIntegrationTests @Autowired constructor(
     }
 
     @Test
-    internal fun `get categories with HttpBearAuth`() {
-        val token = fetchJwt()
+    internal fun `get categories with JWT`() {
+         val token = IntegrationTestUtils.fetchJwt(mockMvc)
 
         mockMvc.get(requestPath) {
             header("Authorization", "Bearer $token")
